# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能知识库平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/knowledge-bases": {
      "get": {
        "tags": [
          "knowledge-bases"
        ],
        "summary": "List Knowledge Bases",
        "operationId": "list_knowledge_bases_api_v1_knowledge_bases_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedData_KnowledgeBaseSummary__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "knowledge-bases"
        ],
        "summary": "Create Knowledge Base",
        "operationId": "create_knowledge_base_api_v1_knowledge_bases_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_KnowledgeBaseSummary_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/knowledge-bases/available": {
      "get": {
        "tags": [
          "knowledge-bases"
        ],
        "summary": "List Available Knowledge Bases",
        "operationId": "list_available_knowledge_bases_api_v1_knowledge_bases_available_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedData_KnowledgeBaseSummary__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/knowledge-bases/{kb_id}": {
      "patch": {
        "tags": [
          "knowledge-bases"
        ],
        "summary": "Update Knowledge Base",
        "operationId": "update_knowledge_base_api_v1_knowledge_bases__kb_id__patch",
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_KnowledgeBaseSummary_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/favorites": {
      "get": {
        "tags": [
          "favorites"
        ],
        "summary": "List Favorites Endpoint",
        "operationId": "list_favorites_endpoint_api_v1_favorites_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Favorites Endpoint Api V1 Favorites Get"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "favorites"
        ],
        "summary": "Create Favorite Endpoint",
        "operationId": "create_favorite_endpoint_api_v1_favorites_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Favorite Endpoint Api V1 Favorites Post"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/favorites/{favorite_id}": {
      "patch": {
        "tags": [
          "favorites"
        ],
        "summary": "Update Favorite Endpoint",
        "operationId": "update_favorite_endpoint_api_v1_favorites__favorite_id__patch",
        "parameters": [
          {
            "name": "favorite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Favorite Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Favorite Endpoint Api V1 Favorites  Favorite Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "favorites"
        ],
        "summary": "Delete Favorite Endpoint",
        "operationId": "delete_favorite_endpoint_api_v1_favorites__favorite_id__delete",
        "parameters": [
          {
            "name": "favorite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Favorite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Favorite Endpoint Api V1 Favorites  Favorite Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_KnowledgeBaseSummary_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeBaseSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "APIResponse[KnowledgeBaseSummary]"
      },
...
      "APIResponse_PaginatedData_KnowledgeBaseSummary__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_KnowledgeBaseSummary_"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "APIResponse[PaginatedData[KnowledgeBaseSummary]]"
      },
...
      "FavoriteCreate": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "answer",
              "document",
              "space",
              "query"
            ],
            "title": "Type"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "maxLength": 4000,
            "title": "Summary",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "title": "Note",
            "default": ""
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Payload"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title"
        ],
        "title": "FavoriteCreate"
      },
      "FavoriteUpdate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "FavoriteUpdate"
      },
...
      "KnowledgeBaseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "chunk_size": {
            "type": "integer",
            "maximum": 4000.0,
            "minimum": 200.0,
            "title": "Chunk Size",
            "default": 800
          },
          "chunk_overlap": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Chunk Overlap",
            "default": 120
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseCreate"
      },
      "KnowledgeBaseSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size"
          },
          "chunk_overlap": {
            "type": "integer",
            "title": "Chunk Overlap"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          },
          "ready_document_count": {
            "type": "integer",
            "title": "Ready Document Count",
            "default": 0
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 0
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "chunk_size",
          "chunk_overlap"
        ],
        "title": "KnowledgeBaseSummary"
      },
      "KnowledgeBaseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(active|archived)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "chunk_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4000.0,
                "minimum": 200.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Size"
          },
          "chunk_overlap": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Overlap"
          }
        },
        "type": "object",
        "title": "KnowledgeBaseUpdate"
      },
...
      "PaginatedData_KnowledgeBaseSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeBaseSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginatedData[KnowledgeBaseSummary]"
      },
...
      "APIErrorResponse": {
        "type": "object",
        "title": "APIErrorResponse",
        "description": "所有非健康检查 API 的统一错误响应。",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API 前缀"
    }
  ]
}